bugfix(radar): Add hero radar objects into its own list to get rid of hero cache updates and its related issues (#1893)

+++ Core/GameEngine/Source/Common/System/Radar.cpp
@@ -81,6 +81,7 @@ void Radar::deleteListResources( void )
 {
 	deleteList(&m_objectList);
 	deleteList(&m_localObjectList);
+	deleteList(&m_localHeroObjectList);
 
 #ifdef DEBUG_CRASHING
 	for( Object *obj = TheGameLogic->getFirstObject(); obj; obj = obj->getNextObject() )
@@ -191,6 +192,7 @@ Radar::Radar( void )
 	m_radarWindow = NULL;
 	m_objectList = NULL;
 	m_localObjectList = NULL;
+	m_localHeroObjectList = NULL;
 	std::fill(m_radarHidden, m_radarHidden + ARRAY_SIZE(m_radarHidden), false);
 	std::fill(m_radarForceOn, m_radarForceOn + ARRAY_SIZE(m_radarForceOn), false);
 	m_terrainAverageZ = 0.0f;
@@ -379,13 +381,13 @@ void Radar::newMap( TerrainLogic *terrain )
 /** Add an object to the radar list.  The object will be sorted in the list to be grouped
 	* using it's radar priority */
 //-------------------------------------------------------------------------------------------------
-RadarObjectType Radar::addObject( Object *obj )
+Bool Radar::addObject( Object *obj )
 {
 
 	// get the radar priority for this object
 	RadarPriorityType newPriority = obj->getRadarPriority();
 	if( isPriorityVisible( newPriority ) == FALSE )
-		return RadarObjectType_None;
+		return FALSE;
 
 	// if this object is on the radar, remove it in favor of the new add
 	RadarObject **list;
@@ -408,26 +410,26 @@ RadarObjectType Radar::addObject( Object *obj )
 	// set a chunk of radar data in the object
 	obj->friend_setRadarData( newObj );
 
-	RadarObjectType objectType;
 	//
 	// we will put this on either the local object list for objects that belong to the
 	// local player, or on the regular object list for all other objects
 	//
 	if( obj->isLocallyControlled() )
 	{
-		list = &m_localObjectList;
-		objectType = RadarObjectType_Local;
+		if ( obj->isHero() )
+			list = &m_localHeroObjectList;
+		else
+			list = &m_localObjectList;
 	}
 	else
 	{
 		list = &m_objectList;
-		objectType = RadarObjectType_Regular;
 	}
 
 	// link object to master list at the head of it's priority section
 	linkRadarObject(newObj, list);
 
-	return objectType;
+	return TRUE;
 }
 
 //-------------------------------------------------------------------------------------------------
@@ -474,24 +476,24 @@ Bool Radar::deleteFromList( Object *obj, RadarObject **list )
 //-------------------------------------------------------------------------------------------------
 /** Remove an object from the radar, the object may reside in any list */
 //-------------------------------------------------------------------------------------------------
-RadarObjectType Radar::removeObject( Object *obj )
+Bool Radar::removeObject( Object *obj )
 {
 
 	// sanity
 	if( obj->friend_getRadarData() == NULL )
-		return RadarObjectType_None;
+		return FALSE;
 
+	if( deleteFromList( obj, &m_localHeroObjectList ) == TRUE )
+		return TRUE;
 	if( deleteFromList( obj, &m_localObjectList ) == TRUE )
-		return RadarObjectType_Local;
+		return TRUE;
 	else if( deleteFromList( obj, &m_objectList ) == TRUE )
-		return RadarObjectType_Regular;
+		return TRUE;
 	else
 	{
-
-		// sanity
 		DEBUG_ASSERTCRASH( 0, ("Radar: Tried to remove object '%s' which was not found",
 											 obj->getTemplate()->getName().str()) );
-		return RadarObjectType_None;
+		return FALSE;
 	}
 
 }
@@ -716,10 +718,14 @@ Object *Radar::objectUnderRadarPixel( const ICoord2D *pixel )
 	// to the radar location
 	//
 
-	// search the local object list
-	obj = searchListForRadarLocationMatch( m_localObjectList, &radar );
+	// search the local hero object list
+	obj = searchListForRadarLocationMatch( m_localHeroObjectList, &radar );
 
-	// search all other objects if not found
+	// search the local object list if not found
+	if( obj == NULL )
+		obj = searchListForRadarLocationMatch( m_localObjectList, &radar );
+
+	// search all other objects if still not found
 	if( obj == NULL )
 		obj = searchListForRadarLocationMatch( m_objectList, &radar );
 
@@ -1359,6 +1365,7 @@ static void xferRadarObjectList( Xfer *xfer, RadarObject **head )
 	* Version Info:
 	* 1: Initial version
 	* 2: TheSuperHackers @tweak Serialize m_radarHidden, m_radarForceOn for each player
+	* 3: TheSuperHackers @tweak Serialize m_localHeroObjectList
 	*/
 // ------------------------------------------------------------------------------------------------
 void Radar::xfer( Xfer *xfer )
@@ -1368,7 +1375,7 @@ void Radar::xfer( Xfer *xfer )
 #if RETAIL_COMPATIBLE_XFER_SAVE
 	XferVersion currentVersion = 1;
 #else
-	XferVersion currentVersion = 2;
+	XferVersion currentVersion = 3;
 #endif
 	XferVersion version = currentVersion;
 	xfer->xferVersion( &version, currentVersion );
@@ -1398,11 +1405,65 @@ void Radar::xfer( Xfer *xfer )
 		xfer->xferUser(&m_radarForceOn, sizeof(m_radarForceOn));
 	}
 
+	if (version <= 2)
+	{
+		if (xfer->getXferMode() == XFER_SAVE)
+		{
+			// TheSuperHackers @info For legacy xfer compatibility.
+			// Transfer all local hero objects to local object list.
+			RadarObject **fromList = &m_localHeroObjectList;
+			RadarObject **toList = &m_localObjectList;
+			while (*fromList != NULL)
+			{
+				RadarObject* nextObject = (*fromList)->friend_getNext();
+				(*fromList)->friend_setNext(NULL);
+				linkRadarObject(*fromList, toList);
+				*fromList = nextObject;
+			}
+		}
+	}
+	else
+	{
+		xferRadarObjectList( xfer, &m_localHeroObjectList );
+	}
+
 	// save our local object list
 	xferRadarObjectList( xfer, &m_localObjectList );
 
 	// save the regular object list
 	xferRadarObjectList( xfer, &m_objectList );
+
+	if (version <= 2)
+	{
+		// TheSuperHackers @info For legacy xfer compatibility.
+		// Transfer hero local object(s) back to local hero object list.
+		// This needs to be done on both load and save.
+		RadarObject **fromList = &m_localObjectList;
+		RadarObject **toList = &m_localHeroObjectList;
+		RadarObject *currObject;
+		RadarObject *prevObject;
+		RadarObject *nextObject;
+		prevObject = NULL;
+		for (currObject = *fromList; currObject != NULL; currObject = nextObject)
+		{
+			nextObject = currObject->friend_getNext();
+			if (currObject->friend_getObject()->isHero())
+			{
+				if (prevObject != NULL)
+				{
+					prevObject->friend_setNext(nextObject);
+				}
+				else
+				{
+					*fromList = nextObject;
+				}
+				currObject->friend_setNext(NULL);
+				linkRadarObject(currObject, toList);
+				continue;
+			}
+			prevObject = currObject;
+		}
+	}
 
 	// save the radar event count and data
 	UnsignedShort eventCountVerify = MAX_RADAR_EVENTS;
